Initial Windows Machine Config Bootstrapper
This commit adds a working bootstrapper, which can be run
on Windows nodes to join them to an existing OpenShift cluster.
The WMCB must be run on the node you wish to join to the cluster,
and requires both a worker ignition file and a kubelet executable
located on the node as well.