Enable winrm when launch Windows instance
This commit Inject PowerShell script to launch Windows instance and open port 5986 for
winrm https inbound traffic so the windows instance is ready for Ansible
communication at launch and every time it boot up.

Change the createInstance() function input with an extra parameter the
contains the user data input, which could be either Windows PowerShell command
or other. So this function is still generic for all the instance type.