Bug 1757446: [wni] Directory validation errors
Cause:
There is no sanity check for whether a path input has content.
There is also a lack of error messages specifying which path input:
kubeconfig, credentials, and windows-node-installer.json file
failed checking against the validation of the path.

Consequence:
The empty path is checked to see if `~` present in the head of the path,
which results in an index out of range error. The error message was
returned directly without knowing which path input is failing.

Fix:
Check if the input path is empty.
Add explicit error messages to places where the validation function
is called.

Result:
Empty input path is errored out with proper messages showing which
input path is not valid.