[wsu] WSU playbook
This commit introduces the Windows Scale Up Playbook. The purpose of
this playbook is to prepare a Windows instance to join an OpenShift
cluster, run the WMCB, and ensure that the node has successfully
joined the cluster. Specifically, this commit implements the
first goal, preparing the instance.

An instance is 'prepared', when it is ready to run the WMCB.
Specifically, when it has a kubelet executable, WMCB executable,
and worker ignition file present on the machine.