[wmcb] Handle optional cloud configuration
The kubelet in an Azure worker node requires a `--cloud-config` command
line argument that specifies the location of a cloud configuration file.
Update WMCB to check if this command line option is present in the worker
ignition's kubelet systemd unit. If it is present, extract it and write
it out to the installation directory and update the Windows kubelet
command line arguments.