[aws-wni] Provide credentials to access Windows Node
As of now, we're not providing information to the user
who ran aws wni to access the Windows VM created, this
commit address that problem. We're now printing the
password to access the Windows VM created along with
the public ipaddress. We're also adding a new top level
package called types, whose intention is to hold all
the structures that we need to use across multiple
cloud providers