[wmcb] test suite to run unit tests This commit builds the unit test binary, transfers the test binary to remote Windows VM created and runs the binary. After that, the test deletes the Windows VM created.
Reason for having an internal directory:

The main test dir shouldn't consider having to spin up a
VM transfer binaries etc. So, I want to have it
in the internal package. This way, the same code
can be reused in WSU and other other places.
Ideally we wouldn't want someone to import
test package from root repo to get framework
which transfers the binaries.