[logging] Allow redirects for Ruby installer download
Currently the module that downloads Ruby installer does not allow redirects,
hence the download does not complete successfully in some cases.
This commit allows the module to accept redirects to solve that problem and adds the
subsequent documentation to the README file.
The added change is using a feature only available in Ansible version 2.9 (the
current latest available), hence the version of Ansible in the README is pinned down
to 2.9.

Bug: 1768568