[logging] E2E tests for logging playbook
This commit add the tests to ensure Ruby and Fluentd are installed correctly
The tests are meant to run on the Windows instance after the logging
playbook has been run successfully against it

The tests added check for the correct version of Ruby installed as well as
the minor version of Fluentd plugin installed.

Ruby installation test depends upon providing the absolute path for the
application. Test written for Ruby Devkit version 2.5.7-x64, install path
assumed to be default.
Fluentd test written for version 1.5.x