Dockerfile for CI
This commit includes a dockerfile which we can use for test running
within OpenShift CI. Documentation for configuring tests in CI
can be found at:
https://github.com/openshift/ci-tools/blob/master/CONFIGURATION.md
Specifcally this dockerfile is planned to be used to build the
"build_root" containers.

This is an attempt to fix the issues we are seeing
with the current CI image, that we cannot install packages in our test
containers, which need ansible and OC. I have opted to not install
ansible, as not all containers need ansible installed, and that can be
done within the individual tests that need it.

This is not guaranteed to work, as the image requirements in the CI docs
are somewhat vague, and its possible that this dockerfile is missing a
CI requirement that is unknown. This can changed in a future commit,
however it is worthwhile to see if this image is usable for our CI jobs
in its current state.