[wsu] Run WMCB through WSU
This commit results in the WSU bootstrapping the specified Windows
nodes, and adding them to the cluster by accepting their CSRs.

In addition, the run-wsu-ci-e2e-test timeout has been increased to 20
minutes. This was done as the runtime of the test was exceeding the
default timeout of 10 minutes and causing test failures.