Upgrade kubelet version to 1.16
Currently WSU playbook downloads kubelet node binary for Kubernetes version 1.14,
however Kubernetes version is upgraded for OpenShift 4.3, from 1.14 to 1.16.
This commit upgrades Kubernetes node binary version to 1.16.2, and also
introduces an e2e test that checks the checksum of binary downloaded to ensure
the binary was downloaded in its entirety.