[wsu] apply worker roles
works only ocp 4.2 or less. In 4.3, the nodeadmission restriction
plugin is blocking nodes to register themselves with worker
labels. So, I am moving the labelling part to WSU. The
eventual goal is to move this to a controller which
is capable of watching for the WMCB label and apply
worker labels. E2e test added to WSU