[wsu] Run the OVN hybrid overlay
This commit results in the WSU downloading and running the OVN hybrid
overlay executable. In addition the commit includes a test to ensure
that the overlay was able to run properly, by checking to make sure the
proper annotation has been added to the node.

The timeout for the WSU e2e tests have also been extended. This is to
compensate for the additional time spent waiting for the hybrid overlay
to run. It has been increased by a fair amount (10 min), so that it
hopefully will not need to be increased in the future.