[wsu] Switch to AMD instance type for e2e tests
This commit switches from using an Intel instance type to using an
AMD instance type when spinning up a VM for our e2e tests.

This is being done as the hybrid overlay executable for Windows,
needed to prepare Windows instances to join a cluster using OVN,
cannot currently be run on Intel instances due to a bug in the
Intel 82599 network driver.