[wni][azure] Fix rule handling
Problem: WNI when creating VMs were overwriting the rules present in the
worker NSG causing due to which we were losing access to the OpenShift
console as the existing rules are required for console access.

Solution: Rather than updating the security group, the code has been
modified to just create / update the rules. To achieve this, the
following was done:
- Remove all instances of updating the worker security group
- Introduce a struct that encapsulates an Azure NSG security rule from a
  WNI perspective
- Add methods to the struct that create/update and delete the rules
- GetMyIP() was exported to allow it be called from the e2e test code