[ci] Improve WMCB unit test failure checking
In addition to checking that "FAIL" is not present in the test output,
add a check that "panic" is not present. We also inspect the exit code
returned by the remote binary. This allows us to catch test failures
that don't print out "FAIL" like "CommandNotFoundException"

To help with debugging we are now printing out the complete output of
the remote test run.