[ci] Allow multiple VMs to be tested at once
This commit refactors the WSU tests, allowing the test suite to test
against multiple VMs at the same time. This resulted in the concept of
tests against a VM and tests against the cluster to be implemented.

I've also added more context to certain test failures, by appending a
list of events related to the pod involved in the failure.