[wsu] Add option to use pinned wmcb
This commit adds the option to use a released WMCB version, by setting
the wmcb_url variable when running the WSU playbook.

I also had to restructure the tests to allow the test suite to tests
against multiple VMs at the same time. This resulted in the concept of
tests against a VM and tests against the cluster to be implemented.

The test timeout was increased again as the scope of the tests have
increased, and more time is required to run them

I've also added more context to certain test failures, by appending a
list of events related to the pod involved in the failure