Bug 1776506: Allow WSU to be run multiple times
This commit ensures that the WSU can be run against the same VM multiple
times without failing.

- Removes the requirement for CSRs to be found. CSRs will only be
present on the initial WSU run.

- Kills existing hybrid-overlay processes before starting a new one, so
that multiples will not exist.

- Removes the kube-proxy service, so that a new kube-proxy executable
can be used if it was changed.

This commit also now has the WSU test suite being run twice against each
VM. Because of that I had to double the test timeout