[wni][azure][aws] Open port 10250 via firewall
Currently we are unable to view container logs via console.
To fix the issue we have opened the port 10250 on windows
node via firewall rule and added a test case in aws & azure
e2e to ensure that. Also added another test in aws e2e which
tests if instance status is ok. Refactored the setup function
in azure e2e to be called before running the actual tests.