Map OpenShift release version to WMCB version
As of now, we're not mapping OpenShift version to WMCB
versions. Without this PR, we'd have to maintain multiple
branches and backport changes. As of now, it doesn't
make sense to maintain branches and backport changes,
in future, we can get started on branching and
picking changes

Logic involved for semver:

curl -s
https://api.github.com/repos/openshift/windows-machine-config-operator/releases
gets the list of all releases, we use the cluster
version from 'oc version' to filter out the release
we need.The wmcb release tags are supposed
to confirm to the following semver format.
v<openshift-major-version>.<openshift-major-version>.\
<wmcb_build_number>-<release_type>.
For example, v4.4-1-alpha.
For a <openshift-major-version-version>.<openshift-minor-version>,
there can be many wmcb builds available, for example,
OpenShift 4.4 can map to v4.4.1-alpha, v4.4.2-alpha,
we will always pick up the latest build available.