[ci] Update Windows VM ami
This commit updates the ami that we use for spinning up a Windows VM in
our wsu and wni e2e tests. The previous ami was removed, so this is a
short term fix to get e2e tests working. A longer term solution would be
finding the latest ami to use, but that will be featured in a future
commit.