[wni][aws][azure] Capture windows user in credentials
This commit ensures that we capture the user information that is used to
access the windows instance. This can be useful when working on the e2e
test cases, especially when interacting with windows VM via winrm client.
Username also varies with cloudprovider, so it would be good to include it
in the credential object.