[wni][aws] Fix subnet selection
Problem: To launch the instance, the first public subnet was being
picked. In the case where the instance type is not supported in the
availability zone the subnet is in, the instance creation would fail.

Solution: Rather picking the first public subnet, we further cross
reference with the list of instance offerings and ensure we pick the
subnet that is in a zone that supports the instance type.

Update the README to highlight usage of AMD instances which is what
surfaced this problem.