[wmcb] Redirect kubelet log to log directory
Currently the kubelet log is generated in kubelet directory (C:\k). This commit
redirects the kublet log to a separate log directory (C:\k\log), which would be
made available in Artifacts directory in the future.

As a part of Unit test, a method for testing creation of Pod Manifest directory
is repurposed to test creation of log directory as well, since both are created
in the same method by bootstrapper.