[wni] Remove log level flag
This commit removes the log level flag from the WNI CLI. This is being
done to prep for the removal of the specific logger currently used,
which will be replaced with a more basic logger. The WNI will not be
supported for direct use in GA, so we do not need full control over
logging levels.