[wni] Bump dependency versions
This commit bumps the kubernetes and OpenShift dependency versions to
kubernetes 1.16.7 and OpenShift 4.3. This is being done to avoid issues
with dependencies when importing the WNI as a library in the WSU test
suite.

This was done by:
* Changing the go.mod file so k8s.io/client-go and k8s.io/apimachinery
  are set to v0.16.7
* Running go get \
  github.com/openshift/api@e9d93e317dd13b6f135baca5cee888d5ff3c222b
  go get \
github.com/openshift/client-go@f6563a70e19a12b2f240eaf4e716ef75baf7003e
  as these are the commits which line up with OpenShift release 4.3
* These dependencies were then placed in the replace section manually,
  so that the entire module is forced to use them.
* Running go mod tidy