[ci] Bump imported WNI to after repo rename
Test builds are failing as the WNI they pull in still reference the old
name for the repository. This commit bumps the imported WNI to the
latest commit to avoid that problem.

This was done by deleting the WNI import line, and then running go mod
tidy. This resulted in go fetching the latest commit of the WNI.