Bug 1805513: Resolve parallel execution of playbook
Problem: When we run wsu playbook parallely for 2 windows vms, the ansible task
'Move wmcb to temporary directory' fails for the second vm. This is because
currently we are building wmcb.exe in the project root directory in both the runs.
So when the run for the first vm has moved the built wmcb.exe to a temporary
directory, run for the second vm cannot find it.

Solution: Instead of moving built wmcb.exe to a temporary directory, we are
copying it