[wni] Refactor WNI to include Windows Interface
As of now, we don't have a way to interact with the windows VM object
we created. This change introduces the Windows VM interface and adds
methods to access to the Windows VM object created using SSH and WinRM.
 Most of the code is moved from framework in the internal package.
This should help in re-using WNI to launch instances in our Windows
Machine Config controller.