[wsu] Capture hybrid-overlay and kube-proxy output
The kube-proxy logs are captured in C:\k\log\kubeProxy\ and the hybrid-overlay
logs in C:\k\log\hybridOverlay\ in the remote VM logs directory.

Kube proxy logs are already being captured so they will just be redirected to
this new directory and hybrid overlay logs are captured using
RedirectStandardError and RedirectStandardOutput arguments in the WSU playbook
while the binaries are running.

Added a method RetrieveDirectories to the WindowsVM interface, this was required
as the remote log directory on the VM can now consist of a nested directory
structure which should be copied to the artifacts directory.

The E2E test makes sure that the log directories and files are being captured in
C:\K\log on the VM.