[wmcb] Redirect logs to stdout
Currently, all WMCB logs are logged in stderr, including info level
logs. This is because the zap package used redirects all logs to stderr
by
(default)[https://godoc.org/sigs.k8s.io/controller-runtime/pkg/log/zap#Options]

This commit sets the info level logs to output to stdout. The logger should
be used for errors, whereas the info and warning level messages should be
output to stdout instead

This is being done because the used version of `zap` package does not
have an option to redirect the logs anywhere else. Moreeover, bumping
its parent package, controller-runtime  to newer versions that does have
that option is not viable because it was simultaneously soft bumped to
go1.13 and kube 1.16