[ci] React to upstream kube-ovn changes
As of now, the ovn-kube network already comes with
hybrid overlay enabled. So, we don't need to patch
the network operator as the cluster comes up with
hybrid overlay enabled. While patching we previously used
to wait till all the nodes have hybrid-overlay specific
annotations. In the new OVN world, the linux nodes
won't have those annotations present, so the test was
failing for us. This commit

- Removes the test which fails if we have don't those annotations
  on master nodes
- Removes patching the network operator object
- Removes the test which checks for the WSU failure when hybrid overlay
  is not enabled