[wni] Update Reinitialize() to handle WinRM client
We are noticing that commands run using Run() that uses a WinRM client
is taking more than 5 minutes to complete after the hybrid-overlay
reconfigures the network. Reinitialize the WinRM client in the
Reinitialize() method to mitigate this.

Add a "warning" pre-fix to the log message printed on errors encountered
when reinitialzing the SSH client to call out that this is not a fatal
error.