[wni] Return if instance is not running
While the AWS instance is getting created, if we manually
delete it or because of some other reason, if the instance
state changes to something other than running, instead of
returning with an error WNI waits till the timeout of
10 minutes and then returns with an error. This is
because of `EC2.GetPasswordData` method which is not
smart enough to get the instance state before checking
for the password. This commit ensures that WNI returns
immediately when the instance terminates.