[wni] Make azure WNI use the given credential file
Previously the credential file argument given to the azure WNI was
thrown away. There was a requirement for an environment variable to be
set and the path given there was used instead.

I have changed the azure e2e suite to create an azure cloud provider
and spin up an instance so that the library is being tested.