Bug 1841140: fix getting kube node binary SHA
This PR fixes two issues encountered while running
WSU with kube version v1.18.3+144c666. The redundant
part in the kube version needs to be trimmed to get the
correct kubernetes version in the grep command.
In addition, since the output of the curl command is no
longer in <td><code>...</code></td> format but in
<td>...</td> format the sha needs to be extracted
accordingly.