Bug 1814989: [docs] Update rate limit section
Update the rate limit section in the WSU README and dev preview docs to
also include the "Get latest 0.8.x cni plugins version" task. This is so
that the user is aware of all the tasks where they could run into GitHub
rate limiting.

Remove the dev preview PDFs as otherwise they would need to be updated
too.