[wni] Expand Cloud interface
As of now, we're using CreateWindowsVM which couples Windows VMs creation along
with attaching it to public subnet. This will not be ideal to us when we are
moving to machine-api in the WMCO world. Machine-api allows the VM to be created
only on the private interface. So, in order to mimick that functionality, this
commit expands the Cloud interface with a method called
CreateWindowsVMWithPrivateIP which allows the WindowsVMs to be created on the
private subnet and won't attach a public IP to the VM created.