[wmcb] Change kubelet log directory
This commit changes the kubelet log directory to C:\var\log\kubelet.
This change is being done to allow `oc adm node-logs` to access the log
file. The --path argument given to that command requires the logs to be
in a sub-directory of /var/log, which resolves to C:\var\log on windows.

WSU changes were made in order to not break existing log collection
functionality.