[WMCB] replace WNI with MachineSets
This commit replaces the usage of WNI with MachineSets.
In order to facilitate this change, we are
1. Replacing KUBECONFIG with in-cluster pod permissions
2. Creating UserDataSecret to instantiate the VMs created using
   MachineSets
3. Removing handleCSR() as it is being done by the
   machine-api-controller

All the required VMs will be created at the same time, we are then
configuring VMs which will be then used for testing.

Also,
this commit makes the switch from winRM to SSH in wmcb tests. Also
removes winRM altogether. SSH is now default to run any command on VM.

Also, updated the powershell script to work with SSH. This was needed as
the current wget powershell script was throwing an error and not
downloading the files as required.