[bootstrapper] Add back Ignition config spec v2.x parsing
With this commit the bootstrapper is now again able to parse
Ignition spec v2 config, and to convert it to spec v3.1 before
further processing.

This dual parsing support is necessary until the windows-mco is
changed to pull Ignition spec v3.1 config from the
machine-config-server (MCS), and can be removed afterwards.