Update Kubelet service in-place
This commit ensures that the idempotency of kubelet service is
mainitained. Recreating the kubelet service caused the bug where
wmcb tried to access kubelet service that was marked for deletion.
The changes include removing deletion of kubelet service, and
ensuring that the kubelet service is created if it is not present,
else updating the existing one with custom specifications.

(cherry picked from commit 8b915fb225c39a05f47bab3df93b77a59e07fa7e)