Refactor to make kubelet args unit-testable
Refactors around the parseIgnitionFileContents() function, in order to
make the generation of kubelet arguments unit-testable. For this commit
I am limiting the scope of this refactor to what I have done here.
There is a lot more refactoring that can be done here, mainly
seperating the file creation and kubelet argument generation into
seperate functions which both take the parsed ignition file as an
input.