[OCPCLOUD-1247] Set correct kubelet hostname for AWS platform
This commit adds a new parameter for the bootstrapper: --platformType.
It allows to define type for the platform where the cluster is
deployed.

If platform type is AWS, then we explicitly override kubelet hostname
with the value from Metadata.

It is required for deploying clusters on Windows machines with
Cloud Controller Manager support, where hostnames must be the same
as node names.

Related links:
https://issues.redhat.com/browse/OCPCLOUD-1247
https://bugzilla.redhat.com/show_bug.cgi?id=1995894