Add 'OpenShift managed' to Kubelet SVC description
The "OpenShift managed" string must be added to the description of all
Windows services created as part of the Windows Node configuration
process. This is a requirement of properly upgrading Nodes. During the
upgrade process, a search can be done for all Windows services with the
string in their description, in order to identify and potentially remove
Services created by us.