Revert addition of dockerRuntime flag
This commit reverts the addition and usage of
dockerRuntime flag option. It was added with
the anticipation of using docker and containerd
runtimes in the same release.
However containerd feature is now expected to
be fully supported in 6.0 timeframe while
removing support for dockerRuntime which makes
this flag option no longer useful.