Remove configure-cni command usage
This commit removes the dev code and tests associated with
running the configure-cni WMCB command. With the removal
of dockerShim flags from kubelet, setting the path for
CNI config and bin directory has been moved to the containerd
config file. Therefore, we no longer need this command to
update the kubelet args with CNI information.
Reference-dockershim flags removal:
[200~https://github.com/kubernetes/kubernetes/pull/106907#issue-1075419