Update containerd and CNI configs
This commit updates the CNI config template
to be in sync with the latest updates on WMCO
side that work with V2 win-overlay API.
It also updates the containerd config to point
to the location of CNI config and bin directory
to the Windows Temp location. The code that was
copying theses files to C:\\k directory was part
of configure CNI test file which seems no longer
necessary to be maintained.

rebase in template